hu4/tests

+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/pragma/r2dbc/entity/EstadoSolicitudEntity.java
@@ -14,7 +14,7 @@ import org.springframework.data.relational.core.mapping.Table;
 public class EstadoSolicitudEntity {
     @Id
     @Column("id_estado" )
-    private short id;
+    private Integer id;
     private String nombre;
     private String descripcion;
 }
